scss review and className fix
