+++ website/src/theme/Footer/LinkItem/index.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import { kebabCase } from 'lodash';
 import Link from '@docusaurus/Link';
 import useBaseUrl from '@docusaurus/useBaseUrl';
 import isInternalUrl from '@docusaurus/isInternalUrl';
 import clsx from 'clsx';
+
 export default function FooterLinkItem({ item }) {
   const { to, href, label, prependBaseUrlToHref, ...props } = item;
   const toUrl = useBaseUrl(to);
@@ -17,7 +19,7 @@ export default function FooterLinkItem({ item }) {
             to: toUrl
           })}
       {...props}
-      className={clsx('footer__link-item', props.className)}>
+      className={clsx('footer__link-item', !!href && `footer__link-item_${kebabCase(item.label)}`)}>
       {label}
       {href && !isInternalUrl(href)}
     </Link>
